requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.31.0
